feat(collection): implement collection display in user profile
refactor: extract components from user profile
