+++ src/api/index.ts
@@ -1,3 +1,4 @@
+export { default as collectionClient } from './collectionClient';
 export { default as problemClient } from './problemClient';
 export { default as recommendClient } from './recommendClient';
 export { default as userClient } from './userClient';
