+++ src/api/mock/index.ts
@@ -1,3 +1,4 @@
+export * from './collections';
 export * from './problems';
 export * from './recommends';
 export * from './user';
