+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as MyCollections } from './MyCollections.vue';
+export { default as MyProfile } from './MyProfile.vue';
