+++ src/utils/config.ts
@@ -1,5 +1,6 @@
 export const { VUE_APP_API_URL, BASE_URL } = process.env;
 
+export const collectionsApiUrl = `${VUE_APP_API_URL}/collections`;
 export const problemsApiUrl = `${VUE_APP_API_URL}/problems`;
 export const recommendsApiUrl = `${VUE_APP_API_URL}/recommends`;
 export const usersApiUrl = `${VUE_APP_API_URL}/users`;
